% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_varImp.R
\name{gg_varImp}
\alias{gg_varImp}
\title{GG Importance Plot}
\usage{
gg_varImp(varImpPlot_matrix, importance_measure = c("\%IncMSE",
  "IncNodePurity"))
}
\arguments{
\item{varImpPlot_matrix}{is the importance matrix of a rf obj.}

\item{importance_measure}{is a string specifying the type of measure to be plotted.}
}
\value{
Returns a ggplot object showing the importance measure.
}
\description{
WIP of Importance Plot for Random Forest models.
}
\details{
Contruct Importance plot from randomForest model.
}
\examples{
set.seed(123)
data(mtcars)

library(randomForest)
mtcars.rf <- randomForest(mpg ~ ., data=mtcars, ntree=1000, keep.forest=FALSE,importance=TRUE)
imp <- importance(mtcars.rf) # save the importance object
gg_varImp(imp, importance_measure = "IncNodePurity")

}
